requirements.txt
# local package
-e .

# external requirements
click~=8.0.4
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas~=1.4.1
setuptools~=57.0.0
scikit-learn~=1.0.2
src~=0.1.0
numpy~=1.21.0
xgboost~=1.5.2
shap~=0.40.0
matplotlib~=3.5.1
seaborn~=0.11.2